deleted listGames.js and add a index for Games to center all information
