+++ Games/index.js
@@ -1,0 +1,11 @@
+
+const nameGamesArray = ["Avalon"];
+
+module.exports = {
+  Avalon : require("./Avalon/avalon-main.js"),
+  nameGames : nameGamesArray
+}
+
+/*
+List of games implemented
+*/
