+++ events/directMessage.js
@@ -1,7 +1,5 @@
 const { Discord, fs, arrayOfFile } = require(`../util/function.js`)
 const config = require(`../config.json`);
-const Games = require(`../listGames.js`);
-
 const mongoose = require("mongoose");
 const { DEFAULTSETTINGS : defaults} = require("../config.json");
 const { Guild } = require("discord.js");
